feat: add modules & env
