+++ src/cli/args.js
@@ -8,9 +8,9 @@ const parseArgs = () => {
   const pairs = process.argv.slice(2);
 
   for (let i = 0; i < pairs.length; i = i + 2) {
-    const propName = pairs[i].slice(2);
+    const prop = pairs[i].slice(2);
     const value = pairs[i + 1];
-    console.log(`${propName} is ${value}`);
+    console.log(`${prop} is ${value}`);
   }
 };
